price/requirement.txt
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-bigquery==2.5.0
google-cloud-core==1.4.4
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
idna==2.10
iexfinance==0.4.3
numpy==1.19.4
pandas==1.1.4
pandas-market-calendars==1.6.1
proto-plus==1.11.0
protobuf==3.14.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
toolz==0.11.1
trading-calendars==2.0.0
urllib3==1.26.2
